Processed @pckroon's comments
